Autentificacion y controllers usuario

+++ backend/controllers/goalController.js
@@ -1,9 +1,11 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 const getGoals = asyncHandler (async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -15,7 +17,8 @@ const setGoals = asyncHandler (async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -29,6 +32,18 @@ const updateGoal = asyncHandler (async (req, res) => {
         throw new Error ('Falta Meta')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('Usuario No Encontrado')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('Usuario No Autorizado')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal)
@@ -40,6 +55,18 @@ const deleteGoal = asyncHandler (async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error ('Falta Meta')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('Usuario No Encontrado')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('Usuario No Autorizado')
     }
 
     await goal.deleteOne()
